Move authentication logic to a new controller mixin
