+++ controller/lib/openshift-origin-controller.rb
@@ -4,6 +4,7 @@ module OpenShift
   module Controller
     require 'controller_engine' if defined?(Rails) && Rails::VERSION::MAJOR == 3
 
+    autoload :Authentication,          'openshift/controller/authentication'
     autoload :ApiBehavior,             'openshift/controller/api_behavior'
     autoload :ApiResponses,            'openshift/controller/api_responses'
     autoload :ActionLog,               'openshift/controller/action_log'
